// src/libraries/formik/types.tsx
export declare type FormikErrors<Values> = {
  [K in keyof Values]?: Values[K] extends any[]
    ? Values[K][number] extends tb.O.Object
      ? FormikErrors<Values[K][number]>[] | string | string[]
      : string | string[]
    : Values[K] extends tb.O.Object
    ? FormikErrors<Values[K]>
    : string
}
/**
 * An object containing touched state of the form whose keys correspond to FormikValues.
 */
export declare type FormikTouched<Values> = {
  [K in keyof Values]?: Values[K] extends any[]
    ? Values[K][number] extends tb.O.Object
      ? FormikTouched<Values[K][number]>[]
      : boolean
    : Values[K] extends tb.O.Object
    ? FormikTouched<Values[K]>
    : boolean
}
/**
 * Formik state tree
 */
export interface FormikState<Values> {
  /** Form values */
  values: Values
  /** map of field names to specific error for that field */
  errors: FormikErrors<Values>
  /** map of field names to whether the field has been touched */
  touched: FormikTouched<Values>
  /** whether the form is currently submitting */
  isSubmitting: boolean
  /** whether the form is currently validating (prior to submission) */
  isValidating: boolean
  /** Top level status state, in case you need it */
  status?: any
  /** Number of times user tried to submit the form */
  submitCount: number
  /** Focused */
  focus?: string
}